Refactor file organization
