+++ src/App.tsx
@@ -4,9 +4,9 @@ import AppBar from './features/AppBar/AppBar';
 import AppFooter from './features/AppFooter/AppFooter';
 import { AuthContext } from './shared/auth-context';
 import { PossibleRoutes } from './utils/constants';
-import MainPage from './pages/MainPage';
-import NewEntryFormPage from './pages/NewEntryFormPage';
-import DashboardPage from './pages/DashboardPage';
+import MainPage from './pages/MainPage/MainPage';
+import NewEntryFormPage from './pages/NewEntryFormPage/NewEntryFormPage';
+import DashboardPage from './pages/DashboardPage/DashboardPage';
 import { ThemeProvider } from '@mui/material/styles';
 import theme from './theme/theme';
 import { initializeApp } from 'firebase/app';
