+++ src/components/JournalEntryCard/JournalEntryCard.css
