+++ src/components/JournalEntryCard/JournalEntryCard.tsx
