+++ src/pages/NewEntryFormPage/NewEntryFormPage.css
